pyproject.toml
[tool.poetry]
name = "codechallenge"
version = "0.1.0"
description = "Just a cool little app"
authors = ["Stefan Padureanu"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.1"
sqlmodel = "^0.0.8"
python-dotenv = "^0.20.0"
uvicorn = "^0.18.3"
asyncpg = "^0.26.0"
alembic = "^1.8.1"
pyjwt = "^2.5.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
